+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add LargestContent Extractor
 * Add KeepEverything Extractor
+* Add NumWordsRules Extractor
 
 # 0.2.0 / 2017-09-11